+++ CHANGELOG.md
@@ -6,8 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 - Compendium items.
 - Set for Charge weapon capability.
+- Aggressive Attack status effect.
 ### Changed
 - Added Set for Charge capability to compendium weapons.
+- Give Ground and Scamper Back toggle off after actor makes an attack.
 
 ## [0.2.18] - 2022-09-26
 ### Added